// prueba-tecnica-04/src/interfaces/interfaces.ts
export interface ProductsTypes {
  id: string
  title: string
  price: number
  rating: number
  reviews_count: number
  image_url: string
  product_url: string
  brand: string
  category: string
}

export interface ProductTypes {
  id: string
  title: string
  price: number
  rating: number
  reviews_count: number
  image_url: string
  product_url: string
  brand: string
  category: string
}